Extracted symbol loading code to it's own file for clarity.
