+++ reference/pageFactory.ts
@@ -1,3 +1,4 @@
+// deno-lint-ignore-file no-explicit-any
 import { ReferenceContext, ReferenceDocumentFactoryFunction } from "./types.ts";
 import getPagesForNamespace from "./_pages/Namespace.tsx";
 import getPagesForNotImplemented from "./_pages/NotImplemented.tsx";
@@ -10,22 +11,35 @@ import getPagesForTypeAlias from "./_pages/TypeAlias.tsx";
 import getPagesForVariable from "./_pages/Variable.tsx";
 import { DocNodeBase } from "@deno/doc/types";
 
-const factories = new Map<string, ReferenceDocumentFactoryFunction<DocNodeBase>>();
-factories.set("moduleDoc", getPagesForModule as ReferenceDocumentFactoryFunction<DocNodeBase>);
-factories.set("namespace", getPagesForNamespace as ReferenceDocumentFactoryFunction<DocNodeBase>);
-factories.set("function", getPagesForFunction as ReferenceDocumentFactoryFunction<DocNodeBase>);
-factories.set("variable", getPagesForVariable as ReferenceDocumentFactoryFunction<DocNodeBase>);
-factories.set("enum", getPagesForNotImplemented as ReferenceDocumentFactoryFunction<DocNodeBase>);
-factories.set("class", getPagesForClass as ReferenceDocumentFactoryFunction<DocNodeBase>);
-factories.set("typeAlias", getPagesForTypeAlias as ReferenceDocumentFactoryFunction<DocNodeBase>);
-factories.set("interface", getPagesForInterface as ReferenceDocumentFactoryFunction<DocNodeBase>);
-factories.set("import", getPagesForImport as ReferenceDocumentFactoryFunction<DocNodeBase>);
+const factories = new Map<
+  string,
+  ReferenceDocumentFactoryFunction<DocNodeBase>
+>();
 
-function factoryFor<T extends DocNodeBase>(item: T): ReferenceDocumentFactoryFunction<T> {
-    return factories.get(item.kind) || getPagesForNotImplemented;
+// Any types used here because denofmt ruins the formatting
+// Safe because this factories collection never escapes this file
+// Actual type: ReferenceDocumentFactoryFunction<DocNodeBase>
+
+factories.set("moduleDoc", getPagesForModule as any);
+factories.set("namespace", getPagesForNamespace as any);
+factories.set("function", getPagesForFunction as any);
+factories.set("variable", getPagesForVariable as any);
+factories.set("enum", getPagesForNotImplemented as any);
+factories.set("class", getPagesForClass as any);
+factories.set("typeAlias", getPagesForTypeAlias as any);
+factories.set("interface", getPagesForInterface as any);
+factories.set("import", getPagesForImport as any);
+
+function factoryFor<T extends DocNodeBase>(
+  item: T,
+): ReferenceDocumentFactoryFunction<T> {
+  return factories.get(item.kind) || getPagesForNotImplemented;
 }
 
-export default function generatePageFor<T extends DocNodeBase>(item: T, context: ReferenceContext) {
-    const factory = factoryFor(item);
-    return factory(item, context) || [];
+export default function generatePageFor<T extends DocNodeBase>(
+  item: T,
+  context: ReferenceContext,
+) {
+  const factory = factoryFor(item);
+  return factory(item, context) || [];
 }
